Fixed types of deltaFree and reservedSpace in FileSystem to 64 bits integers

+++ castor/db/cnv/DbFileSystemCnv.cpp
@@ -527,8 +527,8 @@ void castor::db::cnv::DbFileSystemCnv::createRep(castor::IAddress* address,
     m_insertStatement->setFloat(3, obj->fsDeviation());
     m_insertStatement->setString(4, obj->mountPoint());
     m_insertStatement->setFloat(5, obj->deltaWeight());
-    m_insertStatement->setInt(6, obj->deltaFree());
-    m_insertStatement->setInt(7, obj->reservedSpace());
+    m_insertStatement->setInt64(6, obj->deltaFree());
+    m_insertStatement->setInt64(7, obj->reservedSpace());
     m_insertStatement->setFloat(8, obj->minFreeSpace());
     m_insertStatement->setFloat(9, obj->minAllowedFreeSpace());
     m_insertStatement->setFloat(10, obj->maxFreeSpace());
@@ -597,8 +597,8 @@ void castor::db::cnv::DbFileSystemCnv::updateRep(castor::IAddress* address,
     m_updateStatement->setFloat(3, obj->fsDeviation());
     m_updateStatement->setString(4, obj->mountPoint());
     m_updateStatement->setFloat(5, obj->deltaWeight());
-    m_updateStatement->setInt(6, obj->deltaFree());
-    m_updateStatement->setInt(7, obj->reservedSpace());
+    m_updateStatement->setInt64(6, obj->deltaFree());
+    m_updateStatement->setInt64(7, obj->reservedSpace());
     m_updateStatement->setFloat(8, obj->minFreeSpace());
     m_updateStatement->setFloat(9, obj->minAllowedFreeSpace());
     m_updateStatement->setFloat(10, obj->maxFreeSpace());
@@ -693,8 +693,8 @@ castor::IObject* castor::db::cnv::DbFileSystemCnv::createObj(castor::IAddress* a
     object->setFsDeviation(rset->getFloat(3));
     object->setMountPoint(rset->getString(4));
     object->setDeltaWeight(rset->getFloat(5));
-    object->setDeltaFree(rset->getInt(6));
-    object->setReservedSpace(rset->getInt(7));
+    object->setDeltaFree(rset->getInt64(6));
+    object->setReservedSpace(rset->getInt64(7));
     object->setMinFreeSpace(rset->getFloat(8));
     object->setMinAllowedFreeSpace(rset->getFloat(9));
     object->setMaxFreeSpace(rset->getFloat(10));
@@ -744,8 +744,8 @@ void castor::db::cnv::DbFileSystemCnv::updateObj(castor::IObject* obj)
     object->setFsDeviation(rset->getFloat(3));
     object->setMountPoint(rset->getString(4));
     object->setDeltaWeight(rset->getFloat(5));
-    object->setDeltaFree(rset->getInt(6));
-    object->setReservedSpace(rset->getInt(7));
+    object->setDeltaFree(rset->getInt64(6));
+    object->setReservedSpace(rset->getInt64(7));
     object->setMinFreeSpace(rset->getFloat(8));
     object->setMinAllowedFreeSpace(rset->getFloat(9));
     object->setMaxFreeSpace(rset->getFloat(10));
